pyproject.toml
[project]
name = "ml-zoomcamp-hw"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.120.0",
    "jupyter>=1.1.1",
    "kagglehub>=0.3.13",
    "numpy>=2.3.3",
    "onnx>=1.20.0",
    "onnxruntime>=1.23.2",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "torch>=2.9.1",
    "torchsummary>=1.5.1",
    "torchvision>=0.24.1",
    "uvicorn>=0.38.0",
    "xgboost>=3.1.1",
]
